Add more comments, separate buttons to component
